backend/requirements.txt
django
django-ninja
django-cors-headers
whitenoise
dj-database-url
psycopg2-binary
